--- tests/fixtures/cloudintegrationsutils.py
@@ -1,5 +1,0 @@
-# Deprecation shim: fixtures.cloudintegrationsutils helpers moved into
-# fixtures.cloudintegrations. To be removed once all integration/ import
-# sites are swept to the new path.
-# pylint: disable=wildcard-import,unused-wildcard-import
-from fixtures.cloudintegrations import *  # noqa: F401,F403
